refactoring: TableauDeVariation sort de 2d.js
