+++ src/js/exercices/1e/1E12-2.js
@@ -2,7 +2,7 @@ import Exercice from '../Exercice.js'
 import { listeQuestionsToContenu, numAlpha, randint } from '../../modules/outils.js'
 import Trinome from '../../modules/Trinome.js'
 import { mathalea2d } from '../../modules/2dGeneralites.js'
-import { tableauDeVariation } from '../../modules/2d.js'
+import { tableauDeVariation } from '../../modules/TableauDeVariation'
 export const titre = 'Étude complète de paraboles'
 export const interactifReady = false
