+++ src/js/exercices/1e/1E18.js
@@ -3,8 +3,8 @@ import { combinaisonListes, ecritureAlgebrique, ecritureAlgebriqueSauf1, listeQu
 import Trinome from '../../modules/Trinome.js'
 import FractionX from '../../modules/FractionEtendue.js'
 import { mathalea2d } from '../../modules/2dGeneralites.js'
-import { tableauDeVariation } from '../../modules/2d.js'
 import { context } from '../../modules/context.js'
+import { tableauDeVariation } from '../../modules/TableauDeVariation'
 export const titre = 'Équations et inéquations du second degré'
 export const interactifReady = false
