+++ src/js/exercices/2e/2F31-3.js
@@ -1,7 +1,7 @@
 import Exercice from '../Exercice.js'
 import { mathalea2d } from '../../modules/2dGeneralites.js'
 import { listeQuestionsToContenu, combinaisonListes, randint } from '../../modules/outils.js'
-import { tableauDeVariation } from '../../modules/2d.js'
+import { tableauDeVariation } from '../../modules/TableauDeVariation'
 export const titre = 'Comparaison d\'images dans un tableau de variations'
 
 // Les exports suivants sont optionnels mais au moins la date de publication semble essentielle
