+++ src/js/exercices/Profs/P010.js
@@ -1,7 +1,8 @@
 import Exercice from '../Exercice.js'
 import { mathalea2d } from '../../modules/2dGeneralites.js'
 import { reduirePolynomeDegre3, calcul, texFractionSigne, fractionSimplifiee, listeQuestionsToContenu, printlatex, texNombre, xcas } from '../../modules/outils.js'
-import { tableauDeVariation, repere, courbe, segment, vecteur, rotation, translation, point, tracePoint } from '../../modules/2d.js'
+import { repere, courbe, segment, vecteur, rotation, translation, point, tracePoint } from '../../modules/2d.js'
+import { tableauDeVariation } from '../../modules/TableauDeVariation'
 export const titre = 'Étude de fonctions de degré 3'
 
 /**
