+++ src/js/exercices/can/2e/can2F07.js
@@ -3,8 +3,8 @@ import { mathalea2d } from '../../../modules/2dGeneralites.js'
 import { randint, choice, listeQuestionsToContenu } from '../../../modules/outils.js'
 import { ajouteChampTexteMathLive } from '../../../modules/interactif/questionMathLive.js'
 import { setReponse } from '../../../modules/gestionInteractif.js'
-import { tableauDeVariation } from '../../../modules/2d.js'
 import { context } from '../../../modules/context.js'
+import { tableauDeVariation } from '../../../modules/TableauDeVariation'
 export const titre = 'Lire les extremums dans un tableau de variations'
 export const interactifReady = true
 export const interactifType = 'mathLive'
