+++ src/js/exercices/can/2e/can2F08.js
@@ -3,7 +3,7 @@ import { mathalea2d } from '../../../modules/2dGeneralites.js'
 import { randint, listeQuestionsToContenuSansNumero, sp, choice } from '../../../modules/outils.js'
 import { ajouteChampTexteMathLive } from '../../../modules/interactif/questionMathLive.js'
 import { setReponse } from '../../../modules/gestionInteractif.js'
-import { tableauDeVariation } from '../../../modules/2d.js'
+import { tableauDeVariation } from '../../../modules/TableauDeVariation'
 export const titre = 'Encadrer en utilisant un tableau de variations'
 export const interactifReady = true
 export const interactifType = 'mathLive'
